+++ pkg/message/parser.go
@@ -102,6 +102,9 @@ func convertForeignEncodings(p *parser.Parser) error {
 		RegisterContentTypeHandler("text/.*", func(p *parser.Part) error {
 			return p.ConvertToUTF8()
 		}).
+		RegisterContentTypeHandler("multipart/alternative", func(p *parser.Part) error {
+			return p.ConvertToUTF8()
+		}).
 		RegisterDefaultHandler(func(p *parser.Part) error {
 			t, _, _ := p.Header.ContentType()
 			logrus.WithField("type", t).Trace("Not converting part to utf-8")